rust: say (very non-DRY implementation)
Tried to use as least amount of repetition on strings but ended up
repeating function code a lot more. Could refactor a bit the similar
parts, and see what it looks like. For instance, the last several
functions all are 3 orders of magnitudes different from each other, so I
could see creating a new function that handles this in an exponential
way.

Signed-off-by: Antonio Gutierrez <chibby0ne@gmail.com>