rust: add bracket-push
Used lazy_static crate for initializing a constant HashMap.
Additionally, I took the approach of comparing Option<T> variants
instead of the actual T types (seen in line 37).
Surprisingly enough this solved the bug I had were it was panicking due
to unwrap on `None` variant. This is in contrast to most times were I've
intentionally dealt with the underlying T types and comparing them after
an unwrap.
Will keep a eye out for more of these opportunities.

Signed-off-by: Antonio Gutierrez <chibby0ne@gmail.com>