Add dot-dsl
Needed to use derive macros:
Clone -> For being able to do Vec::from,
Debug -> For printing
PartialEq -> For assert_eq! in tests
Default -> Provides defaults for unspecifed attributes

In addition to that learned about `..self`, which is used by a struct
instantiation as the last member to use the struct's default values,
(i.e: this is only enabled using the Default macro). Very useful when
you don't want to set any attributes and defaults are just fine.

Signed-off-by: Antonio Gutierrez <chibby0ne@gmail.com>