rust: improved acronym from community solutions
Lessons learned:

* skip_while(|p|): skips (i.e: does not yield) the item we are iterating
on **as long as the predicate taken as argument is True**. If the
predicate returns False, skip_while yields the this and all the
following items.

* flat_map(): iterators that works like map but also flattens **one
level** of the nested structure. Very useful when are having to use *a
map operation that instead of returning values *returns Iterators**

Signed-off-by: Antonio Gutierrez <chibby0ne@gmail.com>