rust: add first sieve implementation
Lessons learned:

* Sometimes iterators is a bit too complicated. If there are a lot of
side effects you should think about having an explicit for loop. More
specifically, when the creating an iterator out of a vector or slice and
wanting to remove elements from it, since removing requires a mutable
borrow but creating the iterator was a immutable borrow.

* step_by which is implemented in every iterator. Also the range
expressions implement the Iterator trait and can therefore use
`step_by`.

* map must always return something so if you're doing an if inside the
closure of a map that returns something, it also needs to take an else
that returns the same time. Maps are useful for transforming one type of
iterator into another, e.g: getting rid of the counter when having a
`enumerate()` type of iterator.

* Prototyping in Python and solving a base problem in Pen and Paper
saves a lot of time.

Signed-off-by: Antonio Gutierrez <chibby0ne@gmail.com>