rust/robot-simulator: refactor instructions and use base struct
Instead of repeating:
```
position: self.position
```

use base struct `..self` for the reminder of the fields you don't need
to modify.

Additionally instructions was refactored to use `fold` which allows us
to return a value which is calculated a from a closure and an initial
value. In this case the initail value is the initial robot structure,
and the closure maintains two arguemtns, the last robot structure and
the current letter, which will determine the next robot
position/orientation, which then will be used in another iteration.

This is a perfect use case of `fold` functional priciple.

Signed-off-by: Antonio Gutierrez <chibby0ne@gmail.com>