rust: refactor anagram with mentoring
Use a vector instead of hashmap. Hashmaps have average O(1) time for
get/insert but this constant is big due to calculating the hash. Also
are bigger space wise.

Using a vector requires O(N) time but uses less time and perhaps the
difference in time for smaller words is negligible.

In addition we now use sort of both the chars of the word and the
possible_anagram to compare them, this requires O(N log N) but is the
fastest approach to compare the equality of the strings using vectors.

Signed-off-by: Antonio Gutierrez <chibby0ne@gmail.com>