rust: finally refactor macros correctly
The error about local ambiguity meant that there more than one way to
parse the pattern in any of the tests. This actually was due to the fact
that I had the usual `,` (separator) after the $( key => value)
expression.

But having the optional `$(,)` was enough to satisfy all the test cases,
which were:

* works when nothing is passed
* works with an trailing `,`
* works without a trailing comma
* every 2 key and pair needs to be separated by a comma

Signed-off-by: Antonio Gutierrez <chibby0ne@gmail.com>